requirements.txt
Babel==1.3
Django==1.7.5
Pillow==2.4.0
PyYAML==3.11
Unidecode==0.04.17
argparse==1.2.1
django-appconf==1.0.1
django-compressor==1.4
django-cookie-message==0.1
django-debug-toolbar==1.2.2
django-debug-toolbar-template-timings==0.6.4
django-extensions==1.5.0
django-extra-views==0.6.4
django-haystack==2.3.1
django-json-field==0.5.7
django-jsonfield==0.9.13
django-oauth2-provider==0.2.6.1
django-oscar==1.0.2
django-rest-swagger==0.2.8
django-secure==1.0.1
django-tables2==0.15.0
django-treebeard==2.0
djangorestframework==3.0.5
factory-boy==2.4.1
mock==1.0.1
phonenumbers==6.3.0
purl==1.0.3
pycountry==1.10
python-dateutil==2.4.0
pytz==2014.10
shortuuid==0.4.2
six==1.9.0
sorl-thumbnail==11.12.1b
sqlparse==0.1.14
wsgiref==0.1.2
